refactor: adds page services WITHOUT MIDDLEWARE
