+++ rust-actix/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "mini-x"
-version = "0.1.0"
+name = "waect-rust"
+version = "0.1.1"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
