+++ rust-actix/diesel.toml
@@ -6,4 +6,4 @@ file = "../database/schema.sql"
 custom_type_derives = ["diesel::query_builder::QueryId"]
 
 [migrations_directory]
-dir = "migrations"
+dir = "../migrations"
