+++ rust-actix/src/database/mod.rs
@@ -1,0 +1,3 @@
+pub mod repository;
+pub mod models;
+pub mod schema;
