+++ rust-actix/src/database/models.rs
@@ -1,10 +1,8 @@
-use crate::schema::{followers, messages};
+use crate::database::schema::{followers, messages, users};
 use diesel::prelude::*;
 
-use super::schema::users;
-
 #[derive(Queryable, Selectable)]
-#[diesel(table_name = crate::schema::users)]
+#[diesel(table_name = crate::database::schema::users)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[derive(QueryableByName)]
 pub struct Users {
@@ -23,7 +21,7 @@ pub struct NewUser<'a> {
 }
 
 #[derive(Queryable, Selectable, Insertable)]
-#[diesel(table_name = crate::schema::followers)]
+#[diesel(table_name = crate::database::schema::followers)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[derive(QueryableByName)]
 pub struct Followers {
@@ -39,7 +37,7 @@ pub struct NewFollower<'a> {
 }
 
 #[derive(Queryable, Selectable, Insertable)]
-#[diesel(table_name = crate::schema::messages)]
+#[diesel(table_name = crate::database::schema::messages)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[derive(QueryableByName)]
 pub struct Messages {
