+++ rust-actix/src/frontend/mod.rs
@@ -1,4 +1,4 @@
-pub mod client;
-
+pub mod services;
 pub mod flash_messages;
 pub mod template_structs;
+pub mod routes;
